add delete endpoint for portfolio and programs
