+++ docs/api/apiv3/paths/program.yml
@@ -42,3 +42,85 @@ get:
   description: ''
   operationId: View_program
   summary: View program
+delete:
+  parameters:
+    - description: Program id
+      example: '1'
+      in: path
+      name: id
+      required: true
+      schema:
+        type: integer
+  responses:
+    '204':
+      description: |-
+        Returned if the program was successfully deleted. There is currently no endpoint to query for the actual deletion status.
+        Such an endpoint _might_ be added in the future.
+      headers: {}
+    '403':
+      content:
+        application/hal+json:
+          schema:
+            $ref: "../components/schemas/error_response.yml"
+          examples:
+            response:
+              value:
+                _type: Error
+                errorIdentifier: urn:openproject-org:api:v3:errors:MissingPermission
+                message: You are not authorized to access this resource.
+      description: |-
+        Returned if the client does not have sufficient permissions.
+
+        **Required permission:** admin
+      headers: {}
+    '404':
+      content:
+        application/hal+json:
+          schema:
+            $ref: "../components/schemas/error_response.yml"
+          examples:
+            response:
+              value:
+                _type: Error
+                errorIdentifier: urn:openproject-org:api:v3:errors:NotFound
+                message: The requested resource could not be found.
+      description: |-
+        Returned if the program does not exist or the client does not have sufficient permissions
+        to see it.
+
+        **Required permission:** any permission in the program
+
+        *Note: A client without sufficient permissions shall not be able to test for the existence of
+        a version. That's why a 404 is returned here, even if a 403 might be more appropriate.*
+      headers: {}
+    '406':
+      $ref: "../components/responses/missing_content_type.yml"
+    '415':
+      $ref: "../components/responses/unsupported_media_type.yml"
+    '422':
+      content:
+        application/hal+json:
+          schema:
+            $ref: "../components/schemas/error_response.yml"
+          examples:
+            response:
+              value:
+                _embedded:
+                  details:
+                    attribute: base
+                _type: Error
+                errorIdentifier: urn:openproject-org:api:v3:errors:PropertyConstraintViolation
+                message: Work packages in non descendant projects reference versions
+                  of the program or its descendants.
+      description: |-
+        Returned if the program cannot be deleted. This can happen when there are still references to the program in other workspaces
+        that need to be severed at first.
+      headers: {}
+  tags:
+    - Programs
+  description: |-
+    Deletes the program permanently. As this is a lengthy process, the actual deletion is carried out asynchronously.
+    So the project might exist well after the request has returned successfully. To prevent unwanted changes to
+    the project scheduled for deletion, it is archived at once.
+  operationId: Delete_Program
+  summary: Delete Program
