+++ docs/api/apiv3/paths/project.yml
@@ -71,7 +71,7 @@ delete:
                 message: Work packages in non descendant projects reference versions
                   of the project or its descendants.
       description: |-
-        Returned if the project cannot be deleted. This can happen when there are still references to the project in other projects
+        Returned if the project cannot be deleted. This can happen when there are still references to the project in other workspaces
         that need to be severed at first.
       headers: {}
   tags:
